HAT-37 Правки старта игры

+++ src/main/kotlin/rubber/dutch/hat/app/dto/PlayerDto.kt
@@ -7,7 +7,7 @@ data class PlayerDto(
         val userId: UserId,
         val status: PlayerStatus,
         var moveOrder: Int,
-        var teamId: Long?,
+        var teamId: Int?,
         val role: PlayerRole
 )
